Debug panels moved in separate directory, use consistent naming.
